% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fret.R
\name{calc_fret}
\alias{calc_fret}
\title{Calculate FRET from \code{donor} and \code{acceptor} signals.}
\usage{
calc_fret(donor, acceptor)
}
\arguments{
\item{donor}{Numeric vector of donor signal.}

\item{acceptor}{Numeric vector of acceptor signal.}
}
\value{
Numeric vector, the same length as donor and acceptor.
}
\description{
Calculate FRET from \code{donor} and \code{acceptor} signals.
}
