% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chordNotesByType.R
\name{chordNotesByType}
\alias{chordNotesByType}
\title{Get notes of an arbitrary chord, without regards to underlaying scale/mode.}
\usage{
chordNotesByType(scalenotes, type = "maj", root)
}
\arguments{
\item{scalenotes}{fretNotes object, generated by the fretNotes() function.  A data.frame()
of notes assigned to each fretboard position, numbered and named according
to the underlying scale.}

\item{type}{Character value specifying a particular chord type by abbreviation.
Currently supported types are: "maj", "min", "dim", "aug", "sus2", "sus4",
"maj7", "dom7", "min7", "halfdim7", "dim7", "minmaj7", "aug7".}

\item{root}{Character value specifying the root note of the chord.  If unspecified,
the tonic note (key) of the scale will be used.}
}
\value{
Returns a fretNotes (data.frame) object with the subset of the scalenotes
   argument found in the specified chord, with a "chordnum" column appended,
   plus named attributes "rootnote", "notes" (data.frame), "type" and "name".
}
\description{
Get notes of an arbitrary chord, without regards to underlaying scale/mode.
}
\author{
M.W.Rowe, \email{mike.rowe@gmail.com}
}
