% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chordShapePlots.R
\name{chordShapePlots}
\alias{chordShapePlots}
\title{Plot Major and Minor Chord Shapes by Fretboard Position}
\usage{
chordShapePlots(tonic, pdffile, in.context = TRUE, seven = F)
}
\arguments{
\item{tonic}{Character value specifying root note of chord and tonic of scale.}

\item{pdffile}{Character value specifying path and file name of *.pdf output file.  If
omitted, save plot as "X_chord_shapes.pdf", where X is the root note, with
a "7" appended if seven==TRUE.  Set to NA to prevent a new file from being
opened and closed (to permit external control of graphics device).}

\item{in.context}{Logical; if TRUE, indicate the positions of the other notes in the scale
that are not part of the chord as gray circles.}

\item{seven}{Logical; if TRUE, include the positions of the dominant/minor 7th.}
}
\value{
Returns the notes of the chords as a fretNotes object.
}
\description{
Plot major and minor chord shapes of a given root, individually and within
the neck of a guitar with standard tuning.  Optionally sevenths of each chord
may be included. Six fretboards will be plotted:
\itemize{
  \item All the notes of major pentatonic scale along the neck.
  \item All the major chord tones along the neck.
  \item Major chord diagrams for each of the CAGED chord shapes.
  \item Minor chord diagrams for each of the CAGED chord shapes.
  \item All the minor chord tones along the neck.
  \item All the notes of minor pentatonic scale along the neck.
}
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
