% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diatonicChordNotes.R
\name{diatonicChordNotes}
\alias{diatonicChordNotes}
\title{Find the notes of a diatonic chord within its scale, given the root note}
\usage{
diatonicChordNotes(scalenotes, root = 1, seven = F)
}
\arguments{
\item{scalenotes}{Data.frame() of notes assigned to each fretboard position, numbered and
named according to the underlying scale, such as generated by the
fretNotes() function.}

\item{root}{Integer value specifying the root by number relative to the scale from 1
to 7.}

\item{seven}{Logical; if TRUE, add the 7th of the chord to the triad.}
}
\value{
Returns a fretNotes (data.frame) object with the subset of the scalenotes
   argument found in the specified chord, with a chordnum column appended,
   plus named attributes "rootnote", "notes" (data.frame), "type" and "name".
}
\description{
Find the notes of a diatonic chord within its scale, given the root note
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
