% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawNeck.R
\name{drawNeck}
\alias{drawNeck}
\title{Draw A Fretboard}
\usage{
drawNeck(
  frets,
  fret.space = "between",
  fret.range,
  decorations = TRUE,
  scale.length = 25.5
)
}
\arguments{
\item{frets}{Data.frame returned by fretNotes() function, which will specify numbers
of frets and strings}

\item{fret.space}{Numeric or character value specifying how evenly the frets are spaced;
valid character values are: "accurate" (=12), "even" (=1000) or "between"
(=24).}

\item{fret.range}{2-element numeric vector, specifying range of frets to draw.}

\item{decorations}{Logical; should decorative dots be placed between some frets?}

\item{scale.length}{Numeric value specifying distance from nut to bridge in inches.}
}
\value{
Returns a named vector of the fret positions, with names from "0".
}
\description{
Calculate fret spacing and draw an empty fretboard.
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
