% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawNotes.R
\name{drawNotes}
\alias{drawNotes}
\title{Draw Notes on the Fretboard}
\usage{
drawNotes(
  notes,
  frets.at,
  label = "note",
  cex.pt = 1.6,
  cex.label = 0.6,
  from.fret
)
}
\arguments{
\item{notes}{fretNotes object.}

\item{frets.at}{Numeric}

\item{label}{Character value specifying how to label notes.}

\item{cex.pt}{Numeric value specifying size of note markers.}

\item{cex.label}{Numeric value specifying size of note labels.}

\item{from.fret}{Numeric value specifying how far from the fret the note markers should
be placed.}
}
\value{
Returns TRUE upon completion.
}
\description{
Add note markers and labels at particular locations to a fretboard plot.
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
