% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPositions.R
\name{findPositions}
\alias{findPositions}
\title{Find CAGED Chord Positions}
\usage{
findPositions(scalenotes, min.fret = 0)
}
\arguments{
\item{scalenotes}{fretNotes object, such as generated by the \code{\link{fretNotes}}()
function.  A data.frame() of notes assigned to each fretboard position,
numbered and named according to the underlying scale.}

\item{min.fret}{Integer defining the lowest fret to consider when searching for the first
occurrence of each scale/chord positions.}
}
\value{
Returns a fretNotes object with chord notes
}
\description{
find CAGED chord/scale positions along the fretboard for the present scale.
}
\details{
Position 1 is defined where the tonic note falls under the middle finger
of the 6th string.  In the CAGED system, the tonic chord will use the
E shape.

Positions 2 thru 5 correspond to the CAGED D, C, A and G positions,
respectively.
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
