% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_transitions.R
\name{fret_bin_transitions}
\alias{fret_bin_transitions}
\title{Bin and Summarise States for their Transitions}
\usage{
fret_bin_transitions(data, state, time, frame_time = 0.2)
}
\arguments{
\item{data}{Data frame containing data}

\item{state}{Column that contains state information.}

\item{time}{Column that contains frame information.}

\item{frame_time}{Time per frame.}
}
\value{
\code{tibble()} with summarised transition data.
}
\description{
Bin and Summarise States for their Transitions
}
