% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_calc.R
\name{fret_calc}
\alias{fret_calc}
\title{Calculate FRET from \code{donor} and \code{acceptor} signals.}
\usage{
fret_calc(donor, acceptor)
}
\arguments{
\item{donor}{Numeric vector of donor signal.}

\item{acceptor}{Numeric vector of acceptor signal.}
}
\value{
Numeric vector, the same length as donor and acceptor.
}
\description{
Calculate FRET from \code{donor} and \code{acceptor} signals.
}
