% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_plot.R
\name{fret_plot}
\alias{fret_plot}
\title{Create a FRET-Distance plot from a given R0}
\usage{
fret_plot(R0 = 54, dis_range = 1:120)
}
\arguments{
\item{R0}{R0 For the dye-pair.}

\item{dis_range}{Range of distances in angstroms over which to create the
plot.}
}
\value{
\code{ggplot2} plot.
}
\description{
Create a FRET-Distance plot from a given R0
}
\examples{
library(fretr)
fret_plot(70, seq(150))
}
