% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_plot.R
\name{fret_plot_chrom}
\alias{fret_plot_chrom}
\title{Plot Chromatograms of Raw don & acc and FRET Data}
\usage{
fret_plot_chrom(data, trace_colours = NULL)
}
\arguments{
\item{data}{Data frame containing columns for time, don, acc, and fret}

\item{trace_colours}{Named vector with names of \code{fret}, \code{don} and \code{acc} for
colouring the lines for the traces.}
}
\value{
\code{ggplot} object.
}
\description{
Plot Chromatograms of Raw don & acc and FRET Data
}
