% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fretNotes.R
\name{fretNotes}
\alias{fretNotes}
\title{Assign scale notes to positions of a fretboard}
\usage{
fretNotes(key = "E", scale = "major", n.frets = 22, tuning = c("E",
  "A", "D", "G", "B", "E"))
}
\arguments{
\item{key}{Character value specifying tonic of the scale.}

\item{n.frets}{Integer; number of frets on the fretboard (not counting the nut).}

\item{tuning}{Character vector of length 4 to 8, specifing number and tuning of
strings from lowest to highest by note; default is standard guitar,
c("E","A","D","G","B","E").}

\item{scale:}{Character value specifying scale type; currently "major" or "minor".}
}
\value{
Returns a fretNotes object, a data.frame with one row per fretboard
   position (fret/string combination), and named columns:
   \itemize{
      \item \emph{string}: Integer, 1 to number of strings, high to low.
      \item \emph{fret}: Integer, fret number, where 0 is the nut.
      \item \emph{at}: Number, distance from the nut in inches.
      \item \emph{notenum}: Integer, note number within the scale.
      \item \emph{note}: Character value, note name.
      \item \emph{semitone}: Integer value, interval (in semitones) relative
         to tonic.
   }
   The object will be assigned attributes:
   \itemize{
      \item \emph{key}: character value; tonic of the scale.
      \item \emph{scale}: character value; scale type.
      \item \emph{scalenotes}:
      \item \emph{fretrange}:
   }
}
\description{
Assign scale notes to positions of a fretboard
}
\author{
M.W.Rowe, \email{mike.rowe@gmail.com}
}
