% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelModePlots.R
\name{parallelModePlots}
\alias{parallelModePlots}
\title{Fretboard Maps of All Parallel Modes of a Given Tonic}
\usage{
parallelModePlots(tonic, pdffile)
}
\arguments{
\item{tonic}{A character value specifying the (shared) tonic note of all the modes.}

\item{pdffile}{A character value specifying the path and name of the output *.pdf file
to generate.}
}
\value{
Returns a fretnote object.
}
\description{
Plots seven fretboard maps showing all the modes of a given tonic note,
ordered such that only a single note differs between successive modes.
}
\details{
The modes are ordered according to the cycle of fifths, from brightest to
darkest: lydian, ionian, mixolydian, dorian, aeolian, phrygian, locrian.
The goal is to relate the modes to each other.
}
\author{
M.W.Rowe, \email{mike.rowe@gmail.com}
}
