% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fret.R
\name{plot_fret}
\alias{plot_fret}
\title{Create a FRET-Distance plot from a given R0}
\usage{
plot_fret(R0 = 54, dis_range = 1:120)
}
\arguments{
\item{R0}{R0 For the dye-pair.}

\item{dis_range}{Range of distances in angstroms over which to create the
plot.}
}
\value{
\code{ggplot2} plot.
}
\description{
Create a FRET-Distance plot from a given R0
}
\examples{
library(fretr)
plot_fret(70, seq(150))
}
