% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fret.R
\name{plot_fret_chrom}
\alias{plot_fret_chrom}
\title{Plot Chromatograms of Raw don & acc and FRET Data}
\usage{
plot_fret_chrom(data)
}
\arguments{
\item{data}{Data frame containing columns for time, don, acc, and fret}
}
\value{
\code{ggplot} object.
}
\description{
Plot Chromatograms of Raw don & acc and FRET Data
}
