% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fret.R
\name{plot_tdp}
\alias{plot_tdp}
\title{Create a Transition Density Plot}
\usage{
plot_tdp(data, from, to, nbins = 100, contour = FALSE)
}
\arguments{
\item{data}{Data frame with relevant information.}

\item{from}{Column that contains from_state data.}

\item{to}{Column that contains to_state data.}

\item{nbins}{Number of bins for rasterisation of density.}

\item{contour}{Logical for the inclusion of contours.}
}
\value{
\code{ggplot} object.
}
\description{
Create a Transition Density Plot
}
