% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFretboardByPosition.R
\name{plotFretboardByPosition}
\alias{plotFretboardByPosition}
\title{Plot whole fretboard with positions separated by offset; join frets in common}
\usage{
plotFretboardByPosition(notes, params = list())
}
\arguments{
\item{notes}{A fretNotes object.}

\item{params}{A list object, with named elements specifying various plotting parameters.
It is only necessary to specify those elememts that you wish to change
from their default values:
\itemize{
   \item \strong{shape.column} = "scalenum",
   \item \strong{shapes}
      = c(DtmF="E", mFaFd="D", FdCB="C", CBeFc="A", eFcDt="G"),
}
parameters controlling how fretboards are drawn:
\itemize{
   \item \strong{mar} = c(0.25, 2.5, 1.25, 0.25),
   \item \strong{gap.factor} = 0.8,
   \item \strong{lwd.string} = 1,
   \item \strong{col.string} = "gray85",
   \item \strong{lwd.fret} = 2,
   \item \strong{col.fret} = "gray70",
   \item \strong{lwd.nut} = 5,
   \item \strong{col.nut} = "gray50",
   \item \strong{fret.labels} = c(3, 5, 7, 10, 12, 15, 17, 19, 22),
   \item \strong{cex.fret.labels} = 0.7,
}
parameters controlling how primary notes are drawn:
\itemize{
   \item \strong{note.column} = "scalenum",
   \item \strong{note.label} = "note",
   \item \strong{col.notes} = "black",
   \item \strong{bg.notes} = "blue",
   \item \strong{pch.notes} = 21,
   \item \strong{cex.notes} = 1.7,
   \item \strong{col.text.notes} = "white",
   \item \strong{cex.text.notes} = 0.6,
}
notes to mark but not label (like scale notes around chord notes):
\itemize{
   \item \strong{context.column} = "scalenum",
   \item \strong{col.context} = rgb(0,0,0,0.2),
   \item \strong{bg.context} = rgb(1,1,1,0.2),
   \item \strong{cex.context} = 1.6,
   \item \strong{pch.context} = 21)
}}
}
\value{
Returns frets.
}
\description{
Plot whole fretboard with positions separated by offset; join frets in common
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
