% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_fret.R
\name{pred_fret}
\alias{pred_fret}
\title{Predict the FRET value given a distance and R0.}
\usage{
pred_fret(distance, R0 = 54)
}
\arguments{
\item{distance}{Distance between the two fluorophores.}

\item{R0}{Distance in angstroms at which the two fluorophores have \code{FRET = 0.5}.}
}
\value{
Numeric vector of calculated FRET distances, with the same length as
distance.
}
\description{
Predict the FRET value given a distance and R0.
}
