% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeScalePlots.R
\name{relativeScalePlots}
\alias{relativeScalePlots}
\title{Relative Scale Fretboard Maps (version 2)}
\usage{
relativeScalePlots(tonic, pdffile, params = list())
}
\arguments{
\item{tonic}{A character value specifying the (shared) tonic note of all the modes.}

\item{pdffile}{A character value specifying the path and name of the output *.pdf file
to generate.}

\item{params}{A list.}
}
\description{
Plot fretboards of a scale/mode, and its diatonic chords by position
(alternate version.)
}
\author{
M.W.Rowe, \email{mwr.stats@gmail.com}
}
