% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chordShapePlots.R
\name{chordShapePlots}
\alias{chordShapePlots}
\title{Plot Major and Minor Chord Shapes by Fretboard Position}
\usage{
chordShapePlots(tonic, pdffile, in.context = TRUE, seven = F)
}
\arguments{
\item{tonic}{Character value specifying root note of chord and tonic of scale.}

\item{in.context}{Logical; if TRUE, indicate the positions of the other notes in the scale
that are not part of the chord as gray circles.}

\item{seven}{Logical; if TRUE, include the positions of the dominant/minor 7th.}

\item{pdffile:}{Character value specifying path and file name of *.pdf output file.  If
omitted, save plot as "X_chord_shapes.pdf", where X is the root note, with
a "7" appended if seven==TRUE.  Set to NA to prevent a new file from being
opened and closed (to permit external control of graphics device).}
}
\value{
Returns the notes of the chords as a fretNotes object.
}
\description{
Plot major and minor chord shapes, individually and within neck
}
\author{
M.W.Rowe, \email{mike.rowe@gmail.com}
}
