% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fret_calc.R
\name{fret_calc}
\alias{fret_calc}
\title{Calculate FRET from \code{donor} and \code{acceptor} signals.}
\usage{
fret_calc(donor, trans)
}
\arguments{
\item{donor}{Signal from the donor channel. (Excitation with the donor laser
and emission in donor window.)}

\item{trans}{Signal from the transfer channel. (Excitation with donor laser
and emission in acceptor window.)}
}
\value{
Numeric vector, the same length as donor and acceptor.
}
\description{
Calculate FRET from \code{donor} and \code{acceptor} signals.
}
