% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fretr.R
\docType{package}
\name{fretr}
\alias{fretr}
\title{Generate Fretboard Maps of Scales and Chords}
\description{
This package draws "fretboard maps", positions of notes along the fretboard
of a stringed instrument, such as a guitar.  Typical uses would be to
illustrate all the positions at which a particular chord or scale can be
played, or relationships between scales, modes and chords.
}
\section{Fretboard Map Functions}{

Each of these functions generates a pdf file with sets of fretboards, chord
or scale diagrams meant to illustrate particular relationships between
the scales, chords and their notes.

\itemize{
  \item \code{\link{chordShapePlots}}:
   Plot major and minor chord shapes of a given root, individually and within
   the neck of a guitar with standard tuning.  Optionally sevenths of each
   chord may be included. Six fretboards will be plotted.

  \item \code{\link{scalePositionPlots}}:
     Plots fretboards of a scale/mode, and its diatonic chords by position.
     On the first fretboard, all of the notes of the scale are shown for the
     whole neck. In the second column, the notes of the scale are shown
     broken out by CAGED position on the neck. Next the notes of the major
     pentatonic scale for each position are shown.  The remaining seven
     columns show the CAGED chord shapes at each position, with the dominant
     7th substituted for the diminished chord.

  \item \code{\link{parallelModePlots}}:
     Plots seven fretboard maps showing all the modes of a given tonic note,
     ordered such that only a single note differs between successive modes.
}
}

\section{Low-Level Functions}{

\itemize{
  \item \code{\link{fretNotes}}:
  \item \code{\link{diatonicChordNotes}}
  \item \code{\link{chordNotesByType}}
  \item \code{\link{findPositions}}
  \item \code{\link{makeFretboardString}}
  \item \code{\link{drawNeck}}
  \item \code{\link{plotFretboardByPosition}}
  \item \code{\link{drawNotes}}
}
}

