% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalePositionPlots.R
\name{scalePositionPlots}
\alias{scalePositionPlots}
\title{Fretboard Maps of Diatonic Chords Within a Scale}
\usage{
scalePositionPlots(tonic, scale = "major", pdffile, in.context = FALSE,
  by.5ths = TRUE)
}
\arguments{
\item{tonic}{Character value specifying tonic note of scale or model}

\item{scale}{Character value specifying scale or mode type.}

\item{pdffile}{Character value specifying path and file name of *.pdf output file.  If
omitted, save plot as "X type scale.pdf", where X is the tonic note and
type is the type of scale or mode name.  Set to NA to prevent a new file
from being opened and closed (to permit external control of graphics
device).}

\item{in.context}{Logical; if TRUE, indicate the positions of the other notes in the scale
that are not part of the chord as gray circles.}

\item{by.5ths}{Logical; if TRUE, order chords by the circle of 5ths, starting with the IV
chord.  If FALSE, order by scale degree, I - vii}
}
\value{
Returns a fretNotes objects containing the all notes of the scale or
   mode by fretboard position.
}
\description{
Plot fretboards of a scale/mode, and its diatonic chords by position.
}
\author{
M.W.Rowe, \email{mike.rowe@gmail.com}
}
